do not rely on copy_schedule_dim to find the copy schedule (#12)
copy_schedule_dim is only necessary to construct the space that is used to
extract the proper part of the copy_schedule from the union.  It represents the
number of schedule domains in the range of copy_schedule union, which was
constant before multi-part kernels.  The union contains pw_multi_aff's going
from statement domain spaces to schedule spaces.  Two pw_multi_aff's there
cannot have the same domain as it would be equivalent to having multiple
schedules for the same statement.  Therefore, we can lookup using the domain
space only and thus do not need to know the dimension of the copy_schedule.
Implement this through foreach.