add basic cross band tiling functionality
Cross band tiling takes a tree of the form

         A
         |
         S
     ____|____
     |   |   |
     B  ...  B

and turns it into a tree of the form

         TA
         |
         S
     ____|____
     |   |   |
     TB ...  TB
     |       |
     PA      PA
     PB      PB

This commit introduces functionality for performing the actual tiling,
for checking that the input tree has the right shape and
for checking that the tiling can obviously be applied because
the A node is fully coincident.

Further functionality for checking whether tiling can be applied
when A is not fully coincident will be added in a subsequent commit.