allow indirect promotion when index array is not promoted
In addition to checking whether an array reference group has indirect accesses,
check that the index array reference group was promoted to shared memory. If it
was not, no subscript unwrapping is necessary and the tiling linearization and
index transformation should be performed exactly as for direct accesses.

Die if the index array reference group was promoted to private memory since we
don't expect this to happen.