simplify create_inner_expr
Creation of the piecewise_multi_aff to generate the index array subscript
(mapping AST loops to shared array subscripts) was unnecessarily complicated
and ultimately failed on cases where the original subscripts involved iterators
of loops that were not mapped to blocks or threads.  Replace generation by a
simple map composition assuming an access map is single-valued, that is only a
single element can be accessed.  This assumption is reasonable to make because
we cannot generate a C array subscript that accesses multiple elements at once.