gpu.c: is_candidate: skip domain and context nodes
is_candidate is only called from is_outer_tilable on nodes
starting from the leaves up until a node is found that contains
some permutable bands.  In the default mode of operation,
this never reaches a domain or context node because is_outer_tilable
is only called on trees with some permutable bands.
However, the next commit will allow mapping code to the device
that does not have any permutable bands, in which case
is_candidate can end up getting called on domain and context nodes and
they should not be considered as valid candidates.